Add analysis workflow
This workflow builds the repo on Linux using Gradle and then analyzes the results of the build using Guru Reviewer. For now, this workflow is completely independent from the regular build workflow that Darpan created. Once Guru Reviewer analysis is working correctly, we can unify both workflows, or make one dependent on the other, to avoid the extra build.