Add auto-analysis of full repo using CLI Wrapper (#14)
Same workflow as the one recently added to the Python repo, plus a few additional steps to build the Java code and pass the build artifacts to the CLI.

Changes and news:

* Add a frozen copy of the CLI wrapper that's ready to run.
* Add workflow that uses CLI to analyze whole repo and summarizes results.
* We currently run on every push to any branch (whether in a PR or not).
* Full results are kept as artifacts in the Action run associated with each commit.
* A summary is also printed to the workflow execution log.
* If a PR is open for the branch, the summary is also posted to the PR as a comment.

Adding the 12 MB CLI zip to this repo is a short-term solution. This will improve once the CLI is auto-downloadable by an Action. The CLI has been publicly released, but auto-downloading is pending.